Relocate monitor code into its own module
The env module was doing two separate things and there was enough code
to warrant splitting it up to make each module easier to navigate.
